Fix nat test for test_create_disabled_false
Issues:
Fixes #552

Problem:
When testing against 11.5.4, the JSON being sent to the device had
disabled=False and enabled=True. Only the truthy value should be sent,
thus enabled=True. Even if the user sets disabled=False, we should
modify the resource to ensure the truthy value is set properly and
remove the other Falsy value.

Analysis:
The _endis_able method was previously defined in the nat module and
datacenter module. This has now been moved into a mixin and is avaiable
to any class.

Tests:
The nat tests pass with 11.5.4 and 11.6.0. I will let Jenkins check the
gtm datacenter tests. I also added a few new tests for test_mixins
